refactor: quote folder structure
