+++ src/quote/flashmint/leveraged-extended/provider.ts
@@ -9,9 +9,8 @@ import {
 import { slippageAdjustedTokenAmount } from 'utils/slippage'
 import { Exchange, SwapData } from 'utils'
 
-import { QuoteProvider } from '../quoteProvider'
-import { QuoteToken } from '../quoteToken'
-import { SwapQuoteProvider, SwapQuoteRequest } from '../swap'
+import { QuoteProvider, QuoteToken } from '../../interfaces'
+import { SwapQuoteProvider, SwapQuoteRequest } from '../../swap'
 
 export interface FlashMintLeveragedExtendedQuoteRequest {
   isMinting: boolean
