+++ src/quote/interfaces.ts
@@ -1,3 +1,7 @@
+export interface QuoteProvider<R, Q> {
+  getQuote(request: R): Promise<Q | null>
+}
+
 export interface QuoteToken {
   address: string
   decimals: number
