--- src/quote/quoteProvider.ts
@@ -1,3 +1,0 @@
-export interface QuoteProvider<R, Q> {
-  getQuote(request: R): Promise<Q | null>
-}
